refactor(statistics): split detailed_statistics.py into three focused modules
2312-line monolith split into:
- monthly_statistics_service.py — StatisticsFilters, TypedDicts, monthly
  aggregation, budget data, dashboard summary
- category_statistics_service.py — transaction/receipt filtering, category
  top lists, drilldown, transfer logs
- summary_statistics_service.py — credit cards, charts, balances, main
  get_user_detailed_statistics entry point

detailed_statistics.py kept as a re-export facade so all existing imports
continue to work without changes.
